requirements.txt
# internal
-e .

# external
ratelimit==2.2.1
requests==2.25.1
python-dotenv==0.15.0